Finally got the poison working
It took me so long get a unique poison for the scorps due to some VerbProperties issue. I tried to figure this out by myself for like 2 days, and thought I had to inject C# code. There was no aid online either, but what I did find was a maneuvers,xml file (by changing my file explorer search options to search for specific key works. I searched ToxicBite). Within this file are the "verbs" for moves. From there all you have to do is copy the desired effect and rename it. I had the rest already made a couple days ago (making HeDIff, CapacityDef, CapacityToolDef and Parent for CapacityDef) and now the poison works. Depending on the unit the odds of getting poison varies per strike (roughly 1/3 chance for zealot, and pikeman (if unarmed) and 1/2 chance for shifter). I also added changes to the bodies of the scorpions so it is more inline with reality (i.e me copying scorpion anatomy. I did this anatomy stuff a couple days ago and did not yet test to see if it displays so might change stuff with that). 

Oh ya, the poison is only on tail moves, and pikemen have weapons.

TODO (for now/next steps):
Fix Nudism, balance (raid points), models, custom poisons (king's), and I think that is it (maybe more units).